git-svn-id: https://funk2.googlecode.com/svn/funk2@1340 98e8950a-0dbf-11de-ba4d-252e5d5a8945

+++ trunk/funk2/c/f2_processor_thread_handler.h
@@ -32,7 +32,7 @@ void                      funk2_processor_thread_handler__init(funk2_processor_t
 void                      funk2_processor_thread_handler__destroy(funk2_processor_thread_handler_t* this);
 funk2_processor_thread_t* funk2_processor_thread_handler__add_new_processor_thread(funk2_processor_thread_handler_t* this, funk2_processor_thread_function_pointer_t start_function, void* args);
 
-int this_processor_thread__pool_index();
+u64 this_processor_thread__pool_index();
 
 #endif // F2__PROCESSOR_THREAD_HANDLER__H
